Split source-backed Pro catch-up responsibilities
