+++ crates/ctx-cli/src/semantic/source_backed_pro_catch_up.rs
@@ -1,252 +1,58 @@
-use std::{
-    path::{Path, PathBuf},
-    thread,
-    time::{Duration, Instant},
-};
+#[cfg(test)]
+use std::path::Path;
+use std::time::Duration;
 
-use anyhow::{Context, Result};
-use ctx_history_core::utc_now;
-use ctx_history_index::{
-    release_generation_retention_lease, GenerationRetentionLease, VerifiedIndex,
-};
+use ctx_history_index::VerifiedIndex;
 #[cfg(test)]
 use ctx_pro_host_protocol::CoreMaterializationFinalizationPhase;
+#[cfg(test)]
 use ctx_pro_host_protocol::{
     CoreMaterializationFinalizationPending, CoreMaterializationFinalizationProgress,
     CoreMaterializationReceipt,
 };
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use thiserror::Error;
-
-use crate::{
-    compact_json,
-    pro::{
-        core_finalization_generation_lease, preflight_core_materialization,
-        reconstruct_core_finalization_generation_lease, release_core_finalization_generation_lease,
-        stable_error_code, sync_core_materialization, validate_core_finalization_generation_lease,
-        CoreMaterializationSyncOutcome,
-    },
-};
-
-use super::{
-    paths_status::{
-        daemon_jobs_path, read_daemon_job_status, read_daemon_job_status_strict,
-        write_daemon_job_status,
-    },
-    source_backed_refresh_coordinator::{
-        nonzero_duration_micros, open_verified_index, source_backed_index_root,
-        PinnedCorePublication, PinnedSourceBackedGeneration,
-    },
-};
-
-#[path = "source_backed_pro_catch_up/recheck.rs"]
-mod recheck;
 
 #[cfg(test)]
-use recheck::path as recheck_path;
-pub(super) use recheck::schedule as helper_recheck_schedule;
-use recheck::{
-    complete as complete_observed_recheck, read as read_recheck_request,
-    read_unlocked as read_recheck_request_unlocked, with_lock as with_recheck_lock,
+use crate::pro::{
+    core_finalization_generation_lease, reconstruct_core_finalization_generation_lease,
 };
+
+#[cfg(test)]
+use super::source_backed_refresh_coordinator::{open_verified_index, source_backed_index_root};
+use super::{
+    paths_status::{daemon_jobs_path, write_daemon_job_status},
+    source_backed_refresh_coordinator::{PinnedCorePublication, PinnedSourceBackedGeneration},
+};
+
+mod finalization;
+mod lease_reconciliation;
+mod recheck;
+mod status;
+
+pub(super) use finalization::run_after_core_publication;
+pub(crate) use finalization::wait_for_completed_generation;
+#[cfg(test)]
+use finalization::{
+    run_with, wait_for_completed_generation_with, ProCatchUpAuthority, ProCatchUpSyncOutcome,
+};
+pub(crate) use lease_reconciliation::cancel_core_finalization_generation_lease;
+pub(super) use lease_reconciliation::reconcile_core_finalization_generation_lease;
+pub(super) use recheck::schedule as helper_recheck_schedule;
+#[cfg(test)]
+use recheck::{path as recheck_path, read as read_recheck_request};
 pub(crate) use recheck::{
     publish as publish_helper_recheck_intent, targets as helper_recheck_targets,
     wake as wake_helper_recheck,
 };
+#[cfg(test)]
+use status::{persist_status, status_path, SourceBackedProCatchUpStatus};
+pub(super) use status::{
+    persist_status_json, read_status_json, scheduled_target_generation, status_generation,
+    status_has_finalization_pending,
+};
 
-const SOURCE_BACKED_PRO_CATCH_UP_STATUS_FILE: &str = "pro-catch-up.json";
-const SOURCE_BACKED_PRO_CATCH_UP_SCHEMA_VERSION: u16 = 1;
-const SOURCE_BACKED_PRO_CATCH_UP_POLL_INTERVAL: Duration = Duration::from_millis(50);
-const SOURCE_BACKED_PRO_CATCH_UP_WAIT_TIMEOUT: Duration = Duration::from_secs(60 * 60);
 const SOURCE_BACKED_PRO_CATCH_UP_WAKE_TIMEOUT: Duration = Duration::from_millis(500);
 const SOURCE_BACKED_PRO_CATCH_UP_WAKE_RESPONSE_MAX_BYTES: u64 = 64 * 1024;
-
-#[derive(Debug, Clone, Copy, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum SourceBackedProCatchUpState {
-    Pending,
-    Error,
-    Completed,
-}
-
-#[derive(Debug, Clone, Deserialize, Eq, PartialEq, Serialize)]
-struct SourceBackedProCatchUpStatus {
-    schema_version: u16,
-    owner: String,
-    kind: String,
-    status: SourceBackedProCatchUpState,
-    pending: bool,
-    retryable: bool,
-    core_generation_id: String,
-    receipt_core_generation_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    finalization_progress: Option<CoreMaterializationFinalizationProgress>,
-    attempts: u64,
-    last_attempt_at_ms: i64,
-    #[serde(default)]
-    last_attempt_duration_us: u64,
-    error_code: Option<String>,
-    last_error: Option<String>,
-    #[serde(default)]
-    reason: Option<String>,
-    #[serde(default)]
-    consecutive_failures: u32,
-    #[serde(default)]
-    retry_after_ms: Option<u64>,
-    #[serde(default)]
-    retry_not_before_at_ms: Option<i64>,
-}
-
-impl SourceBackedProCatchUpStatus {
-    fn pending(core_generation_id: &str, attempts: u64) -> Self {
-        Self {
-            schema_version: SOURCE_BACKED_PRO_CATCH_UP_SCHEMA_VERSION,
-            owner: "daemon".to_owned(),
-            kind: "source_backed_pro_catch_up".to_owned(),
-            status: SourceBackedProCatchUpState::Pending,
-            pending: true,
-            retryable: true,
-            core_generation_id: core_generation_id.to_owned(),
-            receipt_core_generation_id: None,
-            finalization_progress: None,
-            attempts,
-            last_attempt_at_ms: utc_now().timestamp_millis(),
-            last_attempt_duration_us: 0,
-            error_code: None,
-            last_error: None,
-            reason: None,
-            consecutive_failures: 0,
-            retry_after_ms: None,
-            retry_not_before_at_ms: None,
-        }
-    }
-
-    fn error(mut self, error: SourceBackedProCatchUpError) -> Self {
-        self.status = SourceBackedProCatchUpState::Error;
-        self.pending = true;
-        self.retryable = error.retryable();
-        if !self.retryable {
-            self.reason = Some(error.code().to_owned());
-        }
-        self.error_code = Some(error.code().to_owned());
-        self.last_error = Some(error.to_string());
-        self
-    }
-
-    fn finalizing(mut self, progress: CoreMaterializationFinalizationProgress) -> Self {
-        self.status = SourceBackedProCatchUpState::Pending;
-        self.pending = true;
-        self.retryable = false;
-        self.receipt_core_generation_id = None;
-        self.finalization_progress = Some(progress);
-        self.error_code = None;
-        self.last_error = None;
-        self.reason = Some("finalizing".to_owned());
-        self.consecutive_failures = 0;
-        self.retry_after_ms = None;
-        self.retry_not_before_at_ms = None;
-        self
-    }
-
-    fn completed(mut self, receipt_generation: String) -> Self {
-        self.status = SourceBackedProCatchUpState::Completed;
-        self.pending = false;
-        self.retryable = false;
-        self.receipt_core_generation_id = Some(receipt_generation);
-        self.finalization_progress = None;
-        self.error_code = None;
-        self.last_error = None;
-        self.reason = None;
-        self.consecutive_failures = 0;
-        self.retry_after_ms = None;
-        self.retry_not_before_at_ms = None;
-        self
-    }
-
-    fn cancelled(mut self, reason: &str) -> Self {
-        self.status = SourceBackedProCatchUpState::Error;
-        self.pending = false;
-        self.retryable = false;
-        self.receipt_core_generation_id = None;
-        self.finalization_progress = None;
-        self.error_code = Some("cancelled".to_owned());
-        self.last_error = Some(reason.to_owned());
-        self.reason = Some("cancelled".to_owned());
-        self.consecutive_failures = 0;
-        self.retry_after_ms = None;
-        self.retry_not_before_at_ms = None;
-        self
-    }
-
-    fn with_duration(mut self, duration_us: u64) -> Self {
-        self.last_attempt_duration_us = duration_us;
-        self
-    }
-
-    fn is_completed_for(&self, core_generation_id: &str) -> bool {
-        self.status == SourceBackedProCatchUpState::Completed
-            && self.core_generation_id == core_generation_id
-            && self.receipt_core_generation_id.as_deref() == Some(core_generation_id)
-    }
-
-    fn is_scheduled_target(&self) -> bool {
-        self.pending
-            && (self.retryable || self.reason.as_deref() == Some("finalizing"))
-            && self.status != SourceBackedProCatchUpState::Completed
-    }
-
-    fn to_json(&self) -> Result<Value> {
-        Ok(compact_json(serde_json::to_value(self)?))
-    }
-}
-
-#[derive(Debug, Error)]
-enum SourceBackedProCatchUpError {
-    #[error(
-        "source_pro_generation_mismatch: expected Core generation {expected}, but {authority} \
-         was supplied by {surface}"
-    )]
-    GenerationMismatch {
-        expected: String,
-        authority: String,
-        surface: &'static str,
-    },
-    #[error("source_pro_index_unavailable: {0}")]
-    IndexUnavailable(String),
-    #[error("{code}: {message}")]
-    Projection { code: String, message: String },
-}
-
-impl SourceBackedProCatchUpError {
-    fn code(&self) -> &str {
-        match self {
-            Self::GenerationMismatch { .. } => "source_pro_generation_mismatch",
-            Self::IndexUnavailable(_) => "source_pro_index_unavailable",
-            Self::Projection { code, .. } => code,
-        }
-    }
-
-    fn projection(error: anyhow::Error) -> Self {
-        Self::Projection {
-            code: stable_error_code(&error)
-                .unwrap_or("pro_core_materialization_unavailable")
-                .to_owned(),
-            message: error.to_string(),
-        }
-    }
-
-    fn retryable(&self) -> bool {
-        match self {
-            Self::GenerationMismatch { .. } => false,
-            Self::IndexUnavailable(_) => true,
-            Self::Projection { code, .. } => !matches!(
-                code.as_str(),
-                "pro_not_installed" | "invalid_response" | "protocol_mismatch" | "cancelled"
-            ),
-        }
-    }
-}
 
 pub(super) struct SourceBackedProCatchUpRun {
     pub(super) status: Value,
@@ -280,511 +86,6 @@ impl<'a> SourceBackedProCoreAuthority<'a> {
             Self::Retained(_) => "retained Core generation pin",
             Self::Durable(_) => "durable active Core generation pin",
         }
-    }
-}
-
-pub(super) fn run_after_core_publication(
-    data_root: &Path,
-    core_generation_id: &str,
-    authority: SourceBackedProCoreAuthority<'_>,
-) -> Result<SourceBackedProCatchUpRun> {
-    if authority.generation_id() != core_generation_id {
-        return record_preflight_error(
-            data_root,
-            core_generation_id,
-            SourceBackedProCatchUpError::GenerationMismatch {
-                expected: core_generation_id.to_owned(),
-                authority: authority.generation_id().to_owned(),
-                surface: authority.surface(),
-            },
-        );
-    }
-    run_with(
-        data_root,
-        core_generation_id,
-        ProCatchUpAuthority {
-            generation_id: Some(authority.generation_id()),
-            verified_index: Some(authority.verified_index()),
-        },
-        preflight_core_materialization,
-        |data_root, index| {
-            Ok(match sync_core_materialization(data_root, index)? {
-                CoreMaterializationSyncOutcome::Finished {
-                    receipt,
-                    did_work,
-                    helper_artifact_sha256,
-                } => ProCatchUpSyncOutcome::Finished {
-                    receipt,
-                    did_work,
-                    helper_artifact_sha256,
-                },
-                CoreMaterializationSyncOutcome::FinalizationPending { pending } => {
-                    ProCatchUpSyncOutcome::FinalizationPending { pending }
-                }
-            })
-        },
-    )
-}
-
-struct ProCatchUpAuthority<'a> {
-    generation_id: Option<&'a str>,
-    verified_index: Option<&'a VerifiedIndex>,
-}
-
-enum ProCatchUpSyncOutcome {
-    Finished {
-        receipt: CoreMaterializationReceipt,
-        did_work: bool,
-        helper_artifact_sha256: String,
-    },
-    FinalizationPending {
-        pending: CoreMaterializationFinalizationPending,
-    },
-}
-
-fn run_with<Preflight, Sync>(
-    data_root: &Path,
-    core_generation_id: &str,
-    authority: ProCatchUpAuthority<'_>,
-    preflight: Preflight,
-    sync: Sync,
-) -> Result<SourceBackedProCatchUpRun>
-where
-    Preflight: FnOnce(&Path) -> Result<()>,
-    Sync: FnOnce(&Path, &VerifiedIndex) -> Result<ProCatchUpSyncOutcome>,
-{
-    // Capture the exact target this run is allowed to satisfy. The completion
-    // path additionally requires the helper identity reported by this exact
-    // materialization session, so an old helper cannot clear a newer intent.
-    let observed_recheck = read_recheck_request(data_root)?;
-    let prior = read_status(data_root);
-    let attempts = next_attempt(prior.as_ref(), core_generation_id);
-    let attempt_started = Instant::now();
-    let pending = prior
-        .as_ref()
-        .filter(|status| status.core_generation_id == core_generation_id)
-        .and_then(|status| status.finalization_progress.clone())
-        .map(|progress| {
-            SourceBackedProCatchUpStatus::pending(core_generation_id, attempts).finalizing(progress)
-        })
-        .unwrap_or_else(|| SourceBackedProCatchUpStatus::pending(core_generation_id, attempts));
-    persist_status(data_root, &pending)?;
-
-    let result: std::result::Result<ProCatchUpSyncOutcome, SourceBackedProCatchUpError> = (|| {
-        if let Some(generation_id) = authority.generation_id {
-            require_generation(
-                core_generation_id,
-                generation_id,
-                "retained Core generation pin",
-            )?;
-        }
-        preflight(data_root).map_err(SourceBackedProCatchUpError::projection)?;
-        let opened_index: VerifiedIndex;
-        let index = match authority.verified_index {
-            Some(index) => index,
-            None => {
-                opened_index = open_exact_index(data_root, core_generation_id)?;
-                &opened_index
-            }
-        };
-        require_generation(
-            core_generation_id,
-            index.generation_id(),
-            "pinned VerifiedIndex",
-        )?;
-        let outcome = sync(data_root, index).map_err(SourceBackedProCatchUpError::projection)?;
-        match &outcome {
-            ProCatchUpSyncOutcome::Finished { receipt, .. } => require_generation(
-                core_generation_id,
-                &receipt.core_generation_id,
-                "Pro Core materialization receipt",
-            )?,
-            ProCatchUpSyncOutcome::FinalizationPending { pending } => {
-                require_generation(
-                    core_generation_id,
-                    &pending.progress.core_generation_id,
-                    "Pro Core finalization progress",
-                )?;
-                // Production acquires this before sending Finish/Continue. This
-                // idempotent reconstruction also covers restart migration and
-                // injected consumers while the exact generation is still pinned.
-                reconstruct_core_finalization_generation_lease(data_root, &pending.progress)
-                    .map_err(SourceBackedProCatchUpError::projection)?;
-            }
-        }
-        Ok(outcome)
-    })();
-
-    match result {
-        Ok(ProCatchUpSyncOutcome::Finished {
-            receipt,
-            did_work,
-            helper_artifact_sha256,
-        }) => {
-            let completed = pending
-                .completed(receipt.core_generation_id)
-                .with_duration(nonzero_duration_micros(attempt_started.elapsed()));
-            persist_status(data_root, &completed)?;
-            // Status is the durable job authority. Publish terminal state before
-            // dropping the target lease so a crash can only retain, never reclaim
-            // an unfinished target.
-            release_core_finalization_generation_lease(data_root, Some(core_generation_id))?;
-            complete_observed_recheck(
-                data_root,
-                observed_recheck.as_ref(),
-                &helper_artifact_sha256,
-            )?;
-            Ok(SourceBackedProCatchUpRun {
-                status: completed.to_json()?,
-                did_work,
-                continuation_pending: false,
-            })
-        }
-        Ok(ProCatchUpSyncOutcome::FinalizationPending {
-            pending: finalization,
-        }) => {
-            let did_work = !finalization.replayed;
-            let finalizing = pending
-                .finalizing(finalization.progress)
-                .with_duration(nonzero_duration_micros(attempt_started.elapsed()));
-            persist_status(data_root, &finalizing)?;
-            Ok(SourceBackedProCatchUpRun {
-                status: finalizing.to_json()?,
-                did_work,
-                continuation_pending: true,
-            })
-        }
-        Err(error) => {
-            let terminal = !error.retryable();
-            let failed = pending
-                .error(error)
-                .with_duration(nonzero_duration_micros(attempt_started.elapsed()));
-            persist_status(data_root, &failed)?;
-            if terminal {
-                release_core_finalization_generation_lease(data_root, Some(core_generation_id))?;
-            }
-            Ok(SourceBackedProCatchUpRun {
-                status: failed.to_json()?,
-                did_work: false,
-                continuation_pending: false,
-            })
-        }
-    }
-}
-
-fn record_preflight_error(
-    data_root: &Path,
-    core_generation_id: &str,
-    error: SourceBackedProCatchUpError,
-) -> Result<SourceBackedProCatchUpRun> {
-    let prior = read_status(data_root);
-    let attempts = next_attempt(prior.as_ref(), core_generation_id);
-    let failed = SourceBackedProCatchUpStatus::pending(core_generation_id, attempts).error(error);
-    persist_status(data_root, &failed)?;
-    Ok(SourceBackedProCatchUpRun {
-        status: failed.to_json()?,
-        did_work: false,
-        continuation_pending: false,
-    })
-}
-
-fn next_attempt(prior: Option<&SourceBackedProCatchUpStatus>, core_generation_id: &str) -> u64 {
-    prior
-        .filter(|status| status.core_generation_id == core_generation_id)
-        .map(|status| status.attempts.saturating_add(1))
-        .unwrap_or(1)
-}
-
-fn require_generation(
-    expected: &str,
-    authority: &str,
-    surface: &'static str,
-) -> std::result::Result<(), SourceBackedProCatchUpError> {
-    if authority == expected {
-        return Ok(());
-    }
-    Err(SourceBackedProCatchUpError::GenerationMismatch {
-        expected: expected.to_owned(),
-        authority: authority.to_owned(),
-        surface,
-    })
-}
-
-fn open_exact_index(
-    data_root: &Path,
-    core_generation_id: &str,
-) -> std::result::Result<VerifiedIndex, SourceBackedProCatchUpError> {
-    let index_root = source_backed_index_root(data_root);
-    let index = open_verified_index(&index_root)
-        .with_context(|| format!("open verified Core index {}", index_root.display()))
-        .map_err(|error| SourceBackedProCatchUpError::IndexUnavailable(format!("{error:#}")))?;
-    require_generation(
-        core_generation_id,
-        index.generation_id(),
-        "pinned VerifiedIndex",
-    )?;
-    Ok(index)
-}
-
-fn status_path(data_root: &Path) -> PathBuf {
-    daemon_jobs_path(data_root).join(SOURCE_BACKED_PRO_CATCH_UP_STATUS_FILE)
-}
-
-fn read_status(data_root: &Path) -> Option<SourceBackedProCatchUpStatus> {
-    read_status_json(data_root).and_then(|value| serde_json::from_value(value).ok())
-}
-
-enum DurableProCatchUpJobRead {
-    Missing,
-    Valid(SourceBackedProCatchUpStatus),
-    Malformed(anyhow::Error),
-}
-
-fn read_durable_status(data_root: &Path) -> DurableProCatchUpJobRead {
-    let value = match read_daemon_job_status_strict(&status_path(data_root)) {
-        Ok(Some(value)) => value,
-        Ok(None) => return DurableProCatchUpJobRead::Missing,
-        Err(error) => {
-            return DurableProCatchUpJobRead::Malformed(anyhow::anyhow!(
-                "invalid_response: durable source-backed Pro catch-up job is unreadable: {error:#}"
-            ))
-        }
-    };
-    match serde_json::from_value(value) {
-        Ok(status) => DurableProCatchUpJobRead::Valid(status),
-        Err(error) => DurableProCatchUpJobRead::Malformed(anyhow::anyhow!(
-            "invalid_response: durable source-backed Pro catch-up job is malformed: {error}"
-        )),
-    }
-}
-
-fn require_durable_status(data_root: &Path) -> Result<Option<SourceBackedProCatchUpStatus>> {
-    match read_durable_status(data_root) {
-        DurableProCatchUpJobRead::Missing => Ok(None),
-        DurableProCatchUpJobRead::Valid(status) => Ok(Some(status)),
-        DurableProCatchUpJobRead::Malformed(error) => Err(error),
-    }
-}
-
-fn persist_status(data_root: &Path, status: &SourceBackedProCatchUpStatus) -> Result<()> {
-    write_daemon_job_status(&status_path(data_root), &status.to_json()?)
-}
-
-pub(super) fn read_status_json(data_root: &Path) -> Option<Value> {
-    read_daemon_job_status(&status_path(data_root))
-}
-
-pub(super) fn persist_status_json(data_root: &Path, status: &Value) -> Result<()> {
-    write_daemon_job_status(&status_path(data_root), status)
-}
-
-pub(super) fn status_generation(data_root: &Path) -> Option<String> {
-    read_status(data_root).map(|status| status.core_generation_id)
-}
-
-pub(super) fn scheduled_target_generation(data_root: &Path) -> Result<Option<String>> {
-    let Some(status) = require_durable_status(data_root)? else {
-        return Ok(None);
-    };
-    if !status.is_scheduled_target() {
-        return Ok(None);
-    }
-    if let Some(progress) = &status.finalization_progress {
-        progress.validate().map_err(|error| {
-            anyhow::anyhow!("invalid durable Pro finalization target: {}", error.message)
-        })?;
-        if progress.core_generation_id != status.core_generation_id {
-            anyhow::bail!(
-                "invalid durable Pro finalization target: progress generation does not match its job"
-            );
-        }
-    }
-    Ok(Some(status.core_generation_id))
-}
-
-/// Reconciles the single durable Core-generation lease with the daemon's
-/// durable Pro job before the scheduler attempts to pin that job's target.
-pub(super) fn reconcile_core_finalization_generation_lease(data_root: &Path) -> Result<()> {
-    let lease = core_finalization_generation_lease(data_root)?;
-    let Some(status) = require_durable_status(data_root)? else {
-        if lease.is_some() {
-            cancel_core_finalization_generation_lease(
-                data_root,
-                "stale Pro finalization lease had no durable job",
-            )?;
-        }
-        return Ok(());
-    };
-    if status.schema_version != SOURCE_BACKED_PRO_CATCH_UP_SCHEMA_VERSION
-        || status.owner != "daemon"
-        || status.kind != "source_backed_pro_catch_up"
-    {
-        if lease.is_some() {
-            cancel_core_finalization_generation_lease(
-                data_root,
-                "durable Pro finalization job identity was invalid",
-            )?;
-            return Ok(());
-        }
-        anyhow::bail!(
-            "invalid_response: durable source-backed Pro catch-up job identity is invalid"
-        );
-    }
-
-    let terminal = !status.pending
-        || status.status == SourceBackedProCatchUpState::Completed
-        || (status.status == SourceBackedProCatchUpState::Error && !status.retryable);
-    if terminal {
-        if let Some(lease) = lease {
-            release_observed_generation_lease(data_root, &lease)?;
-        }
-        return Ok(());
-    }
-
-    if let Some(progress) = &status.finalization_progress {
-        let validation = (|| -> Result<()> {
-            progress
-                .validate()
-                .map_err(|error| anyhow::anyhow!("invalid_response: {}", error.message))?;
-            if progress.core_generation_id != status.core_generation_id {
-                anyhow::bail!(
-                    "invalid_response: durable Pro finalization progress targets a foreign Core generation"
-                );
-            }
-            if lease.is_some() {
-                validate_core_finalization_generation_lease(data_root, progress)
-            } else {
-                reconstruct_core_finalization_generation_lease(data_root, progress)
-            }
-        })();
-        if let Err(error) = validation {
-            if lease.is_some() {
-                cancel_core_finalization_generation_lease(data_root, &error.to_string())?;
-            } else {
-                let failed = status
-                    .clone()
-                    .error(SourceBackedProCatchUpError::projection(error));
-                persist_status(data_root, &failed)?;
-            }
-        }
-        return Ok(());
-    }
-
-    let Some(lease) = lease else {
-        return Ok(());
-    };
-    if lease.generation_id() != status.core_generation_id {
-        cancel_core_finalization_generation_lease(
-            data_root,
-            "durable Core generation lease targeted a foreign Pro job",
-        )?;
-        return Ok(());
-    }
-    if status.pending && status.status != SourceBackedProCatchUpState::Completed {
-        // Finish can commit while its response is lost. Until helper status
-        // supplies the finalization tuple, the generation-bound pending job is
-        // the only durable identity available and the bounded lease must remain.
-        return Ok(());
-    }
-    release_core_finalization_generation_lease(data_root, Some(&status.core_generation_id))?;
-    Ok(())
-}
-
-pub(crate) fn cancel_core_finalization_generation_lease(
-    data_root: &Path,
-    reason: &str,
-) -> Result<bool> {
-    let Some(lease) = core_finalization_generation_lease(data_root)? else {
-        return Ok(false);
-    };
-    let attempts = require_durable_status(data_root)?
-        .as_ref()
-        .filter(|status| status.core_generation_id == lease.generation_id())
-        .map(|status| status.attempts)
-        .unwrap_or(1);
-    let cancelled =
-        SourceBackedProCatchUpStatus::pending(lease.generation_id(), attempts).cancelled(reason);
-    persist_status(data_root, &cancelled)?;
-    release_observed_generation_lease(data_root, &lease)
-}
-
-fn release_observed_generation_lease(
-    data_root: &Path,
-    lease: &GenerationRetentionLease,
-) -> Result<bool> {
-    release_generation_retention_lease(source_backed_index_root(data_root), lease).map_err(|error| {
-        anyhow::anyhow!(
-            "invalid_response: durable Core generation lease changed before terminal release: {error}"
-        )
-    })
-}
-
-pub(super) fn status_has_finalization_pending(data_root: &Path, core_generation_id: &str) -> bool {
-    read_status(data_root).is_some_and(|status| {
-        status.status == SourceBackedProCatchUpState::Pending
-            && status.pending
-            && !status.retryable
-            && status.reason.as_deref() == Some("finalizing")
-            && status.core_generation_id == core_generation_id
-            && status
-                .finalization_progress
-                .as_ref()
-                .is_some_and(|progress| progress.core_generation_id == core_generation_id)
-    })
-}
-
-/// Waits for the daemon-owned Pro projection of one exact Core generation.
-///
-/// This is deliberately a status-only seam. The CLI must not reread provider
-/// sources or materialize Pro itself; that work remains owned by the daemon
-/// tick that retained the generation-bound Core authority.
-pub(crate) fn wait_for_completed_generation(
-    data_root: &Path,
-    core_generation_id: &str,
-) -> Result<()> {
-    wait_for_completed_generation_with(
-        data_root,
-        core_generation_id,
-        SOURCE_BACKED_PRO_CATCH_UP_WAIT_TIMEOUT,
-        || thread::sleep(SOURCE_BACKED_PRO_CATCH_UP_POLL_INTERVAL),
-    )
-}
-
-fn wait_for_completed_generation_with(
-    data_root: &Path,
-    core_generation_id: &str,
-    timeout: Duration,
-    mut wait: impl FnMut(),
-) -> Result<()> {
-    let started = Instant::now();
-    loop {
-        let (pending_recheck, status) = with_recheck_lock(data_root, || {
-            Ok((
-                read_recheck_request_unlocked(data_root)?.is_some(),
-                read_status(data_root),
-            ))
-        })?;
-        if let Some(status) = status {
-            if !pending_recheck && status.is_completed_for(core_generation_id) {
-                return Ok(());
-            }
-            if status.core_generation_id == core_generation_id
-                && status.status == SourceBackedProCatchUpState::Error
-            {
-                let code = status.error_code.as_deref().unwrap_or("not_materialized");
-                let detail = status
-                    .last_error
-                    .as_deref()
-                    .unwrap_or("daemon source-backed Pro catch-up failed");
-                anyhow::bail!("{code}: {detail}");
-            }
-        }
-        if started.elapsed() >= timeout {
-            anyhow::bail!(
-                "not_materialized: timed out waiting for daemon source-backed Pro generation {core_generation_id}"
-            );
-        }
-        wait();
     }
 }
 
@@ -1719,7 +1020,14 @@ mod tests {
 
     #[test]
     fn production_catch_up_has_no_manifest_resolver_or_provider_io() {
-        let source = include_str!("source_backed_pro_catch_up.rs");
+        let source = [
+            include_str!("source_backed_pro_catch_up.rs"),
+            include_str!("source_backed_pro_catch_up/finalization.rs"),
+            include_str!("source_backed_pro_catch_up/lease_reconciliation.rs"),
+            include_str!("source_backed_pro_catch_up/recheck.rs"),
+            include_str!("source_backed_pro_catch_up/status.rs"),
+        ]
+        .join("\n");
         for forbidden in [
             ["Source", "Manifest"].concat(),
             ["source", "_manifest"].concat(),
